Struct twilight_http_ratelimiting::in_memory::InMemoryRatelimiter

pub struct InMemoryRatelimiter { /* private fields */ }

Default ratelimiter implementation used in twilight that stores ratelimit information in an in-memory mapping.

This will meet most users’ needs for simple ratelimiting, but for multi-processed bots, consider either implementing your own Ratelimiter that uses a shared storage backend or use the HTTP proxy.

Implementations

impl InMemoryRatelimiter

pub fn new() -> Self

Create a new in-memory ratelimiter.

This is used by HTTP client to queue requests in order to avoid hitting the API’s ratelimits.