Module twilight_http_ratelimiting::request

Request parameters for ratelimiting.

This module contains the type definitions for parameters relevant for ratelimiting.

The super::Ratelimiter uses Paths and Methods to store and associate buckets with routes.

Structs

PathParseError

Error returned when a Path could not be parsed from a string.

Enums

Method

Request method.

Path

An enum representing a path, most useful for ratelimiting implementations.

PathParseErrorType

Type of PathParseError specifying what failed to parse.