Struct twilight_http_ratelimiting::Bucket

pub struct Bucket { /* private fields */ }

A bucket containing ratelimiting information for a Path.

Implementations

impl Bucket

pub const fn new( limit: u64, remaining: u64, reset_after: Duration, started_at: Option<Instant>, ) -> Self

Create a representation of a ratelimiter bucket.

Buckets are returned by ratelimiters via Ratelimiter::bucket method. Its primary use is for informational purposes, including information such as the number of remaining tickets or determining how much time remains until the bucket interval resets.

pub const fn limit(&self) -> u64

Total number of tickets allotted in a cycle.

pub const fn remaining(&self) -> u64

Number of tickets remaining.

pub const fn reset_after(&self) -> Duration

Duration after the Self::started_at time the bucket will refresh.

pub const fn started_at(&self) -> Option<Instant>

When the bucket’s ratelimit refresh countdown started.

pub fn time_remaining(&self) -> Option<Duration>

How long until the bucket will refresh.

May return None if the refresh timer has not been started yet or the bucket has already refreshed.